+++ Lora-Bot/Models/BinaryPacket.cs
@@ -6,23 +6,7 @@ using Fraunhofer.Fit.IoT.Bots.LoraBot.Parser;
 
 namespace Fraunhofer.Fit.IoT.Bots.LoraBot.Models {
   public class BinaryPacket : Packet {
-    public BinaryPacket(String name, LoraParser.Typ typ, Single lat, Single lon, Single hdop, Single height, Single battery, RecievedData recieveddata) {
-      this.TransferData = new LocationUpdateEvent(typ) {
-        Name = name,
-        BatteryLevel = battery,
-        Gps = new GpsUpdateEvent() {
-          Latitude = lat,
-          Longitude = lon,
-          Hdop = hdop,
-          Height = height,
-          Fix = lat != 0 || lon != 0
-        }
-      };
-      this.Type = typ;
-      this.SetLoraData(recieveddata, this.TransferData);
-    }
-
-    public BinaryPacket(String name, LoraParser.Typ typ, Single lat, Single lon, Single hdop, Single height, Single battery, UInt16 counter, Boolean has_time, Boolean has_date, Boolean has_fix, Byte sat, Boolean correct_if, String hash, RecievedData recieveddata) {
+    public BinaryPacket(String name, LoraParser.Typ typ, Single lat, Single lon, Double hdop, Double height, Double battery, UInt16 counter, Boolean has_time, Boolean has_date, Boolean has_fix, Byte sat, Boolean correct_if, String hash, RecievedData recieveddata) {
       this.TransferData = new LocationUpdateEvent(typ) {
         Name = name,
         BatteryLevel = battery,